refactor(stella-tui): the ISSUES tab's selection modals move beside its other keys
`deck_ui.rs` is a god file closed to growth, and this branch adds a mode to
`IssuesMode`, a field to `IssuesPanel` and two dispatch lines to it. Another
PR grew the same file by nine lines; each was under its own ceiling and the
two compose over it.

So the send-to-prompt confirmation moves to `issues_keys.rs`, where the
`issues_prompt_text` it calls already lived and where the start-work handler
this branch adds now sits. The split it leaves is a real one: the two modal
surfaces that hold a *selection* are in the key module, and the three that
hold a text buffer stay beside the buffers they type into.

Refs #5044
